Move resin hole activate to resin hole system
